requirements.txt
paho-mqtt==1.6.1
numpy>=1.26.0
pandas>=2.1.0
streamlit>=1.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
